// src/app/model/InformePago.ts
export class InformePago{
    collectionId: string;
    collectionStatus: string;
    paymentId: string;
    status: string;
    externalReference: string;
    paymentType: string;
    merchantOrderId: string;
    preferenceId: string;
    siteId: string;
    processingMode: string;
    reservaAutomatica: boolean;
}